source-mysql-batch: Move discovery into separate file
